# modules/gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bedrock Gateway API
  description: |
    Multi-tenant SaaS proxy for Amazon Bedrock with Cognito authentication,
    cascading budget management, rate limiting, and an admin dashboard.

    ## Authentication

    All API endpoints require authentication via one of:
    - **Cognito JWT Bearer Token**: `Authorization: Bearer <access_token>`
    - **Legacy Token Exchange**: `POST /auth/exchange` (deprecated)

    Human users authenticate via Cognito PKCE flow using the `bg-cognito-auth.sh` CLI.
    Agents/services use Cognito `client_credentials` flow for M2M authentication.

    ## Rate Limiting

    Rate limits are enforced at organization, department, team, and user levels.
    When rate limited, responses include:
    - `Retry-After` header
    - `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset` headers

    ## Budget Enforcement

    Budget limits cascade from organization to user level.
    When budget is exceeded (hard mode), requests are rejected with 429.
    Response headers include: `X-Budget-Remaining-USD`, `X-Budget-Period`
  version: 1.0.0
  contact:
    name: Platform Team
    email: platform@company.com
  license:
    name: Private
    url: https://company.com/license
servers:
  - url: https://{cloudfront_domain}/api
    description: Production Gateway via CloudFront
    variables:
      cloudfront_domain:
        default: gateway.company.com
        description: CloudFront distribution domain
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/exchange:
    post:
      tags: [authentication]
      summary: "[DEPRECATED] Exchange AWS credentials for gateway token"
      description: |
        **DEPRECATED**: Use Cognito JWT authentication instead.
        - Human users: Use PKCE flow with bg-cognito-auth.sh
        - Agents/Services: Use client_credentials flow with Cognito App Client
      deprecated: true
      operationId: exchangeCredentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthExchangeRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthExchangeResponse'
        '401':
          description: Invalid AWS credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied - unknown organization or unregistered service account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [authentication]
      summary: Get current user information
      description: Returns information about the currently authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: User information returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          description: Invalid or expired token
...
  /auth/revoke:
    post:
      tags: [authentication]
      summary: Revoke current token
      operationId: revokeToken
      responses:
        '200':
          description: Token revoked successfully
        '401':
          description: Invalid or expired token
  /auth/service-accounts:
    post:
      tags: [authentication]
      summary: Create a new service account
      description: Register an IAM role as a service account for M2M authentication
      operationId: createServiceAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceAccountCreate'
      responses:
        '201':
          description: Service account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAccountResponse'
        '403':
          description: Insufficient permissions
        '409':
          description: Service account already exists
    get:
      tags: [authentication]
      summary: List service accounts
      operationId: listServiceAccounts
      parameters:
        - name: department_id
          in: query
          schema:
            type: string
        - name: team_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of service accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAccountListResponse'
  /auth/service-accounts/{service_account_id}:
    get:
      tags: [authentication]
      summary: Get service account details
      operationId: getServiceAccount
      parameters:
        - name: service_account_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAccountResponse'
        '404':
          description: Service account not found
    put:
      tags: [authentication]
      summary: Update service account
      operationId: updateServiceAccount
      parameters:
        - name: service_account_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceAccountUpdate'
      responses:
        '200':
          description: Service account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAccountResponse'
    delete:
      tags: [authentication]
      summary: Delete service account
      operationId: deleteServiceAccount
      parameters:
        - name: service_account_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service account deleted
...
  /auth/admin/revoke-user-tokens/{user_id}:
    post:
      tags: [authentication]
      summary: Revoke all tokens for a user
      operationId: revokeAllUserTokens
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tokens revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tokens_revoked:
                    type: integer

  # ==========================================================================
  # Proxy Endpoints
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito JWT access token
  schemas:
...
    AuthExchangeRequest:
      type: object
      required:
        - aws_access_key_id
        - aws_secret_access_key
        - aws_session_token
      properties:
        aws_access_key_id:
          type: string
        aws_secret_access_key:
          type: string
        aws_session_token:
          type: string
    AuthExchangeResponse:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        user_id:
          type: string
        org_id:
          type: string
        team_id:
          type: string
        department_id:
          type: string
        account_type:
          type: string
          enum: [human, service]
    CurrentUserResponse:
      type: object
      properties:
        user_id:
          type: string
        org_id:
          type: string
        department_id:
          type: string
        team_id:
          type: string
        account_type:
          type: string
        is_admin:
          type: boolean
        expires_at:
          type: string
          format: date-time
    ServiceAccountCreate:
      type: object
      required:
        - name
        - department_id
        - team_id
        - iam_role_arn
      properties:
        name:
          type: string
          maxLength: 255
        department_id:
          type: string
        team_id:
          type: string
        iam_role_arn:
          type: string
          maxLength: 512
    ServiceAccountUpdate:
      type: object
      properties:
        name:
          type: string
        department_id:
          type: string
        team_id:
          type: string
        iam_role_arn:
          type: string
    ServiceAccountResponse:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        department_id:
          type: string
        team_id:
          type: string
        iam_role_arn:
          type: string
        created_at:
          type: string
          format: date-time
    ServiceAccountListResponse:
      type: object
      properties:
        service_accounts:
          type: array
          items:
            $ref: '#/components/schemas/ServiceAccountResponse'
        total_count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer

    # Proxy Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object